create repo when using pack cache
